Importance sampling
Instead of bouncing rays into random directions,
sample a new direction from a distribution that is similar
to the BRDF itself. This means we are more likely to calculate
samples that have a higher impact on the final output, reducing variance.

There are also some slight tweaks to light calculations.